Display Post Titles within Component
This is a commit mostly to prove we have a full-stack app.  We are
making an api request upon page load that communicates to the server to
grab all the posts, then carry that back up to the state where they can
be displayed on any page.  Our site does that now, and it is small and
mighty.