Add Translation route
Added a new component for the Translation Proof of Concept.  To keep the
working spaces clean I added a new route on our App component, and will
work only on that route.  I also cleaned up some styling on the App
component to have better distinction from the modules we are importing
versus the directory specific components we are importing.