Add animation for translation on hover
I added a basic animation to show what it would look like to hover over
a sentence and have it translate, and not have that experience be mad
unpleasant.  Now there's some blurry text that settles into the
translated text.  I am not sure if it is mad unpleasant, but it is at least a
little softer than before.